First cut for reorganizing shared code
