+++ container_service_extension/config_validator.py
@@ -44,6 +44,7 @@ from container_service_extension.server_constants import SYSTEM_ORG_NAME
 from container_service_extension.server_constants import \
     VCENTER_LOGIN_ERROR_MSG
 from container_service_extension.server_constants import VERSION_V1
+from container_service_extension.server_utils import should_use_mqtt_protocol
 from container_service_extension.telemetry.telemetry_utils import\
     store_telemetry_settings
 from container_service_extension.uaaclient.uaaclient import UaaClient
@@ -51,7 +52,6 @@ from container_service_extension.utils import check_file_permissions
 from container_service_extension.utils import check_keys_and_value_types
 from container_service_extension.utils import get_duplicate_items_in_list
 from container_service_extension.utils import NullPrinter
-from container_service_extension.utils import should_use_mqtt_protocol
 from container_service_extension.utils import str_to_bool
