+++ container_service_extension/def_/utils.py
@@ -1,96 +1,12 @@
 # container-service-extension
 # Copyright (c) 2020 VMware, Inc. All Rights Reserved.
 # SPDX-License-Identifier: BSD-2-Clause
-from enum import Enum
-from enum import unique
+
+"""Utility methods to help interaction with defined entities framework."""
 
 from container_service_extension.cloudapi.cloudapi_client import CloudApiClient
+import container_service_extension.def_.constants as def_constants
 import container_service_extension.exceptions as excptn
-
-# Defined Entity Framework related constants
-DEF_CSE_VENDOR = 'cse'
-DEF_VMWARE_VENDOR = 'vmware'
-DEF_VMWARE_INTERFACE_NSS = 'k8s'
-DEF_VMWARE_INTERFACE_VERSION = '1.0.0'
-DEF_VMWARE_INTERFACE_NAME = 'Kubernetes'
-DEF_TKG_ENTITY_TYPE_NSS = 'tkgcluster'
-DEF_TKG_ENTITY_TYPE_VERSION = '1.0.0'
-DEF_INTERFACE_ID_PREFIX = 'urn:vcloud:interface'
-DEF_NATIVE_ENTITY_TYPE_NSS = 'nativeCluster'
-DEF_NATIVE_ENTITY_TYPE_VERSION = '1.0.0'
-DEF_NATIVE_ENTITY_TYPE_NAME = 'nativeClusterEntityType'
-DEF_NATIVE_ENTITY_TYPE_RIGHT_BUNDLE = \
-    f'{DEF_CSE_VENDOR}:{DEF_NATIVE_ENTITY_TYPE_NSS} Entitlement'
-
-DEF_ENTITY_TYPE_ID_PREFIX = 'urn:vcloud:type'
-DEF_API_MIN_VERSION = 35.0
-DEF_SCHEMA_DIRECTORY = 'cse_def_schema'
-DEF_ENTITY_TYPE_SCHEMA_FILE = 'schema.json'
-DEF_ERROR_MESSAGE_KEY = 'message'
-DEF_RESOLVED_STATE = 'RESOLVED'
-TKG_ENTITY_TYPE_NSS = 'tkgcluster'
-TKG_ENTITY_TYPE_VERSION = '1.0.0'
-
-
-@unique
-class DefKey(str, Enum):
-    INTERFACE_VENDOR = 'interface_vendor'
-    INTERFACE_NSS = 'interface_nss'
-    INTERFACE_VERSION = 'interface_version'
-    INTERFACE_NAME = 'interface_name'
-    ENTITY_TYPE_VENDOR = 'entity_type_vendor'
-    ENTITY_TYPE_NAME = 'entity_type_name'
-    ENTITY_TYPE_NSS = 'entity_type_nss'
-    ENTITY_TYPE_VERSION = 'entity_type_version'
-    ENTITY_TYPE_SCHEMA_VERSION = 'schema_version'
-
-
-MAP_API_VERSION_TO_KEYS = {
-    35.0: {
-        DefKey.INTERFACE_VENDOR: DEF_VMWARE_VENDOR,
-        DefKey.INTERFACE_NSS: DEF_VMWARE_INTERFACE_NSS,
-        DefKey.INTERFACE_VERSION: DEF_VMWARE_INTERFACE_VERSION,
-        DefKey.INTERFACE_NAME: DEF_VMWARE_INTERFACE_NAME,
-        DefKey.ENTITY_TYPE_VENDOR: DEF_CSE_VENDOR,
-        DefKey.ENTITY_TYPE_NSS: DEF_NATIVE_ENTITY_TYPE_NSS,
-        DefKey.ENTITY_TYPE_VERSION: DEF_NATIVE_ENTITY_TYPE_VERSION,
-        DefKey.ENTITY_TYPE_NAME: DEF_NATIVE_ENTITY_TYPE_NAME,
-        DefKey.ENTITY_TYPE_SCHEMA_VERSION: 'api_v35',
-    },
-    36.0: {
-        DefKey.INTERFACE_VENDOR: DEF_VMWARE_VENDOR,
-        DefKey.INTERFACE_NSS: DEF_VMWARE_INTERFACE_NSS,
-        DefKey.INTERFACE_VERSION: DEF_VMWARE_INTERFACE_VERSION,
-        DefKey.INTERFACE_NAME: DEF_VMWARE_INTERFACE_NAME,
-        DefKey.ENTITY_TYPE_VENDOR: DEF_CSE_VENDOR,
-        DefKey.ENTITY_TYPE_NSS: DEF_NATIVE_ENTITY_TYPE_NSS,
-        DefKey.ENTITY_TYPE_VERSION: DEF_NATIVE_ENTITY_TYPE_VERSION,
-        DefKey.ENTITY_TYPE_NAME: DEF_NATIVE_ENTITY_TYPE_NAME,
-        DefKey.ENTITY_TYPE_SCHEMA_VERSION: 'api_v35',
-    }
-}
-
-
-class ClusterEntityFilterKey(Enum):
-    """Keys to filter cluster entities in CSE (or) vCD.
-
-    Below Keys are commonly used filters. An entity can be filtered by any of
-    its properties.
-
-    Usage examples:
-    ..api/cse/internal/clusters?entity.kind=native
-    ..api/cse/internal/clusters?entity.metadata.org_name=org1
-    ..cloudapi/1.0.0/entities?filter=entity.metadata.org_name==org1
-    """
-
-    # TODO(DEF) CLI can leverage this enum for the filter implementation.
-    CLUSTER_NAME = 'name'
-    ORG_NAME = 'entity.metadata.org_name'
-    OVDC_NAME = 'entity.metadata.ovdc_name'
-    KIND = 'entity.kind'
-    K8_DISTRIBUTION = 'entity.spec.k8_distribution.template_name'
-    STATE = 'state'
-    PHASE = 'entity.status.phase'
 
 
 def raise_error_if_def_not_supported(cloudapi_client: CloudApiClient):
@@ -98,7 +14,8 @@ def raise_error_if_def_not_supported(cloudapi_client: CloudApiClient):
 
     :param cloudapi_client CloudApiClient
     """
-    if float(cloudapi_client.get_api_version()) < DEF_API_MIN_VERSION:
+    if float(cloudapi_client.get_api_version()) < \
+            def_constants.DEF_API_MIN_VERSION:
         raise excptn.DefNotSupportedException("Defined entity framework is not"
                                               " supported for {cloudapi_client.get_api_version()}")  # noqa: E501
 
@@ -127,7 +44,7 @@ def generate_interface_id(vendor, nss, version):
 
     :rtype str
     """
-    return f"{DEF_INTERFACE_ID_PREFIX}:{vendor}:{nss}:{version}"
+    return f"{def_constants.DEF_INTERFACE_ID_PREFIX}:{vendor}:{nss}:{version}"
 
 
 def generate_entity_type_id(vendor, nss, version):
@@ -142,4 +59,4 @@ def generate_entity_type_id(vendor, nss, version):
 
     :rtype str
     """
-    return f"{DEF_ENTITY_TYPE_ID_PREFIX}:{vendor}:{nss}:{version}"
+    return f"{def_constants.DEF_ENTITY_TYPE_ID_PREFIX}:{vendor}:{nss}:{version}"  # noqa: E501
