+++ container_service_extension/server_constants.py
@@ -2,6 +2,9 @@
 # Copyright (c) 2019 VMware, Inc. All Rights Reserved.
 # SPDX-License-Identifier: BSD-2-Clause
 
+"""Constants used only in the CSE server."""
+
+from dataclasses import dataclass
 from enum import Enum
 from enum import unique
 
@@ -85,9 +88,47 @@ REQUEST_ID_FORMAT = 'Request Id: %(requestId)s | '
 
 
 @unique
+class OperationType(str, Enum):
+    CLUSTER = 'cluster'
+    NODE = 'node'
+    OVDC = 'ovdc'
+    SYSTEM = 'system'
+    TEMPLATE = 'template'
+
+
+@unique
 class ThreadLocalData(str, Enum):
     USER_AGENT = 'User-Agent'
     REQUEST_ID = 'request_id'
+
+
+@unique
+class DefEntityOperation(str, Enum):
+    CREATE = 'CREATE'
+    DELETE = 'DELETE'
+    UPDATE = 'UPDATE'
+    UPGRADE = 'UPGRADE'
+    UNKNOWN = 'UNKNOWN'
+
+
+@unique
+class DefEntityOperationStatus(str, Enum):
+    IN_PROGRESS = 'IN_PROGRESS'
+    SUCCEEDED = 'SUCCEEDED'
+    FAILED = 'FAILED'
+    UNKNOWN = 'UNKNOWN'
+
+
+@unique
+class FlattenedClusterSpecKey(Enum):
+    WORKERS_COUNT = 'workers.count'
+    NFS_COUNT = 'nfs.count'
+    TEMPLATE_NAME = 'k8_distribution.template_name'
+    TEMPLATE_REVISION = 'k8_distribution.template_revision'
+
+
+VALID_UPDATE_FIELDS = [FlattenedClusterSpecKey.WORKERS_COUNT.value, FlattenedClusterSpecKey.NFS_COUNT.value,  # noqa: E501
+                       FlattenedClusterSpecKey.TEMPLATE_NAME.value, FlattenedClusterSpecKey.TEMPLATE_REVISION.value]  # noqa: E501
 
 
 CLUSTER_ENTITY = 'cluster_entity'
@@ -456,3 +497,38 @@ CSE_SERVICE_ROLE_RIGHTS = [
     "vmware:tkgcluster: Modify",
     "vmware:tkgcluster: View"
 ]
+
+
+@dataclass
+class DefEntityPhase:
+    """Supports two ways of creation.
+
+    1. DefEntityPhase(DefEntityOperation.CREATE, DefEntityOperationStatus.SUCCEEDED) # noqa: E501
+    2. DefEntityPhase.from_phase('CREATE:SUCCEEDED')
+    """
+
+    operation: DefEntityOperation
+    status: DefEntityOperationStatus
+
+    def __str__(self):
+        return f'{self.operation}:{self.status}'
+
+    @classmethod
+    def from_phase(cls, phase: str):
+        """Return instance of DefEntityPhase.
+
+        :param str phase: defined entity phase value. ex: "CREATE:SUCCEEDED"
+        :return: DefEntityPhase
+        :rtype: <class DefEntityPhase>
+        """
+        operation, status = phase.split(':')
+        return cls(DefEntityOperation[operation], DefEntityOperationStatus[status])  # noqa: E501
+
+    def is_operation_status_success(self) -> bool:
+        try:
+            return self.status == DefEntityOperationStatus.SUCCEEDED
+        except Exception:
+            return False
+
+    def is_entity_busy(self) -> bool:
+        return self.status == DefEntityOperationStatus.IN_PROGRESS
